LRDOCS-8703 Rename *Greeting classes to *Greeter
