+++ docs/dxp/7.x/en/liferay-internals/fundamentals/apis-as-osgi-services/resources/liferay-p9g2.zip/p9g2-api/src/main/java/com/acme/p9g2/Greeter.java
@@ -1,9 +1,9 @@
-package com.acme.j1h1;
+package com.acme.p9g2;
 
 import aQute.bnd.annotation.ProviderType;
 
 @ProviderType
-public interface Greeting {
+public interface Greeter {
 
 	public void greet(String name);
