+++ docs/dxp/7.x/en/liferay-internals/fundamentals/apis-as-osgi-services/resources/liferay-p9g2.zip/p9g2-impl/src/main/java/com/acme/p9g2/internal/P9G2Greeter.java
@@ -1,11 +1,11 @@
 package com.acme.p9g2.internal;
 
-import com.acme.p9g2.Greeting;
+import com.acme.p9g2.Greeter;
 
 import org.osgi.service.component.annotations.Component;
 
-@Component(service = Greeting.class)
-public class P9G2Greeting implements Greeting {
+@Component(service = Greeter.class)
+public class P9G2Greeter implements Greeter {
 
 	@Override
 	public void greet(String name) {
