+++ docs/dxp/7.x/en/liferay-internals/fundamentals/using-an-osgi-service/resources/liferay-j1h1.zip/j1h1-api/src/main/java/com/acme/j1h1/Greeter.java
@@ -1,9 +1,9 @@
-package com.acme.k8s2;
+package com.acme.j1h1;
 
 import aQute.bnd.annotation.ProviderType;
 
 @ProviderType
-public interface Greeting {
+public interface Greeter {
 
 	public void greet(String name);
