+++ docs/dxp/7.x/en/liferay-internals/fundamentals/using-an-osgi-service/resources/liferay-j1h1.zip/j1h1-impl/src/main/java/com/acme/j1h1/internal/J1H1Greeter.java
@@ -1,11 +1,11 @@
 package com.acme.j1h1.internal;
 
-import com.acme.j1h1.Greeting;
+import com.acme.j1h1.Greeter;
 
 import org.osgi.service.component.annotations.Component;
 
-@Component(service = Greeting.class)
-public class J1H1Greeting implements Greeting {
+@Component(service = Greeter.class)
+public class J1H1Greeter implements Greeter {
 
 	@Override
 	public void greet(String name) {
